+++ common/src/main/kotlin/com/cobbledex/config/CobbleDexConfig.kt
@@ -11,6 +11,7 @@ data class CobbleDexConfig(
     val showObtainment: Boolean = true,
     val showDrops: Boolean = true,
     val showStats: Boolean = true,
+    val showMoves: Boolean = true,
     val showPokedexInfo: Boolean = true,
     val localDatapackScan: Boolean = true,
     val debugMode: Boolean = false