+++ common/src/main/kotlin/com/cobbledex/jei/CobbleDexJEIPlugin.kt
@@ -10,6 +10,8 @@ import com.cobbledex.jei.drops.JeiDropCategory
 import com.cobbledex.jei.drops.JeiDropRecipe
 import com.cobbledex.jei.evolution.JeiEvolutionCategory
 import com.cobbledex.jei.evolution.JeiEvolutionRecipe
+import com.cobbledex.jei.moves.JeiMovesCategory
+import com.cobbledex.jei.moves.JeiMovesRecipe
 import com.cobbledex.jei.obtainment.JeiObtainmentCategory
 import com.cobbledex.jei.obtainment.JeiObtainmentRecipe
 import com.cobbledex.jei.pokedex.JeiPokedexInfoCategory
@@ -49,6 +51,7 @@ open class CobbleDexJEIPlugin : IModPlugin {
         if (config.showObtainment) registration.addRecipeCategories(JeiObtainmentCategory(guiHelper))
         if (config.showDrops) registration.addRecipeCategories(JeiDropCategory(guiHelper))
         if (config.showStats) registration.addRecipeCategories(JeiStatsCategory(guiHelper))
+        if (config.showMoves) registration.addRecipeCategories(JeiMovesCategory(guiHelper))
         if (config.showPokedexInfo) registration.addRecipeCategories(JeiPokedexInfoCategory(guiHelper))
         DebugLog.info("JEI: Categories registered")
     }
@@ -89,6 +92,12 @@ open class CobbleDexJEIPlugin : IModPlugin {
             DebugLog.info("JEI: Registered ${statsRecipes.size} stats recipes")
         }
 
+        if (config.showMoves) {
+            val movesRecipes = RecipeBuilder.buildAllMovesRecipes().map { JeiMovesRecipe(it) }
+            registration.addRecipes(JeiMovesCategory.RECIPE_TYPE, movesRecipes)
+            DebugLog.info("JEI: Registered ${movesRecipes.size} moves recipes")
+        }
+
         if (config.showPokedexInfo) {
             val pokedexRecipes = RecipeBuilder.buildAllPokedexInfoRecipes().map { JeiPokedexInfoRecipe(it) }
             registration.addRecipes(JeiPokedexInfoCategory.RECIPE_TYPE, pokedexRecipes)
@@ -103,6 +112,7 @@ open class CobbleDexJEIPlugin : IModPlugin {
         if (config.showObtainment) registration.addRecipeCatalyst(ItemStack(Items.NETHER_STAR), JeiObtainmentCategory.RECIPE_TYPE)
         if (config.showDrops) registration.addRecipeCatalyst(ItemStack(Items.DIAMOND), JeiDropCategory.RECIPE_TYPE)
         if (config.showStats) registration.addRecipeCatalyst(ItemStack(Items.BOOK), JeiStatsCategory.RECIPE_TYPE)
+        if (config.showMoves) registration.addRecipeCatalyst(ItemStack(Items.PAPER), JeiMovesCategory.RECIPE_TYPE)
         if (config.showPokedexInfo) registration.addRecipeCatalyst(ItemStack(Items.WRITABLE_BOOK), JeiPokedexInfoCategory.RECIPE_TYPE)
     }
 }