+++ common/src/main/kotlin/com/cobbledex/jei/moves/JeiMovesRecipe.kt
@@ -1,0 +1,7 @@
+package com.cobbledex.jei.moves
+
+import com.cobbledex.MovesRecipeData
+
+data class JeiMovesRecipe(val data: MovesRecipeData) {
+    val speciesName get() = data.speciesName
+}