+++ common/src/main/kotlin/com/cobbledex/rei/moves/MovesDisplay.kt
@@ -1,0 +1,37 @@
+package com.cobbledex.rei.moves
+
+import com.cobbledex.CobbleDexMod
+import com.cobbledex.MovesRecipeData
+import com.cobbledex.sanitizePath
+import com.cobbledex.rei.entry.PokemonEntry
+import com.cobbledex.rei.entry.PokemonEntryType
+import me.shedaniel.rei.api.common.category.CategoryIdentifier
+import me.shedaniel.rei.api.common.display.Display
+import me.shedaniel.rei.api.common.entry.EntryIngredient
+import me.shedaniel.rei.api.common.entry.EntryStack
+import net.minecraft.resources.ResourceLocation
+import java.util.Optional
+
+class MovesDisplay(val data: MovesRecipeData) : Display {
+
+    val speciesName get() = data.speciesName
+
+    private val cachedInputEntries: List<EntryIngredient> by lazy {
+        listOf(EntryIngredient.of(EntryStack.of(PokemonEntryType.POKEMON, PokemonEntry(speciesName))))
+    }
+
+    override fun getInputEntries(): List<EntryIngredient> = cachedInputEntries
+
+    override fun getOutputEntries(): List<EntryIngredient> = emptyList()
+
+    override fun getCategoryIdentifier(): CategoryIdentifier<*> = MovesCategory.ID
+
+    override fun getDisplayLocation(): Optional<ResourceLocation> {
+        return Optional.of(
+            ResourceLocation.fromNamespaceAndPath(
+                CobbleDexMod.MOD_ID,
+                "moves/${sanitizePath(speciesName)}_${data.pageIndex}"
+            )
+        )
+    }
+}